requirements_dev.txt
pip==22.3.1
mysql.connector==2.2.9
datetime==5.1